feat(syntax): add syntax mod

+++ compiler/ac_ir/src/source.rs
@@ -1,5 +1,5 @@
 #[salsa::tracked(debug)]
 pub struct FileId<'db> {
     #[tracked]
-    pub value: usize,
+    pub value: u32,
 }
